fixed linux build errors

+++ libs/api/TcpServer.cpp
@@ -10,7 +10,7 @@ static void echo_alloc(uv_handle_t* handle,
 	size_t suggested_size,
 	uv_buf_t* buf) {
 	buf->base = (char*)malloc(suggested_size);
-	buf->len = (ULONG)suggested_size;
+	buf->len = suggested_size;
 }
 
 static void close_cb(uv_handle_t* handle) {
